Fix PR feedback
